Mon, Mar 25, 2024, 11:19 AM +01:00

+++ calculator.c
@@ -9,7 +9,7 @@ int main(void)
 {
     int x = get_int("x: ");
     int y = get_int("y: ");
-    float z = (float) x / y;
+    float z = (float) x / (float) y;
 
     for(char i[20]= ""; strlen(i) < 10; strcat(i, "r"))
     {
